+++ src/Slices/ServiceInventory/UI/Components/SetStateAction.tsx
@@ -1,20 +1,17 @@
 import React, { useContext, useState } from "react";
 import {
-  Alert,
-  AlertActionCloseButton,
-  AlertGroup,
-  Button,
   Dropdown,
   DropdownItem,
   DropdownToggle,
-  Modal,
-  ModalVariant,
+  Text,
 } from "@patternfly/react-core";
 import { CaretDownIcon } from "@patternfly/react-icons";
 import { Maybe, VersionedServiceInstanceIdentifier } from "@/Core";
 import { ActionDisabledTooltip } from "@/UI/Components";
 import { DependencyContext } from "@/UI/Dependency";
 import { words } from "@/UI/words";
+import ConfirmationModal from "./ConfirmationModal";
+import { ToastAlertMessage } from "./ToastAlertMessage";
 
 interface Props extends VersionedServiceInstanceIdentifier {
   targets: string[] | null;
@@ -106,101 +103,16 @@ export const SetStateAction: React.FC<Props> = ({
         />
       </ActionDisabledTooltip>
       <ConfirmationModal
-        confirmationText={confirmationText}
+        title={words("inventory.statustab.confirmTitle")}
         onSetInstanceState={onSubmit}
         id={id}
         targetState={targetState}
         isModalOpen={isModalOpen}
         setIsModalOpen={handleModalToggle}
         setErrorMessage={setStateErrorMessage}
-      />
-    </>
-  );
-};
-
-interface ModalProps {
-  confirmationText: string;
-  id: string;
-  targetState: string;
-  onSetInstanceState: (targetState: string) => Promise<void>;
-  isModalOpen: boolean;
-  setIsModalOpen: (isOpen: boolean) => void;
-  setErrorMessage: (message: string) => void;
-}
-
-const ConfirmationModal: React.FC<ModalProps> = ({
-  confirmationText,
-  onSetInstanceState,
-  isModalOpen,
-  setIsModalOpen,
-  id,
-  targetState,
-}) => {
-  const handleModalToggle = () => {
-    setIsModalOpen(!isModalOpen);
-  };
-  const handleConfirm = async () => {
-    handleModalToggle();
-    await onSetInstanceState(targetState);
-  };
-
-  return (
-    <React.Fragment>
-      <Modal
-        data-testid={`${id}-set-state-modal`}
-        variant={ModalVariant.small}
-        title={words("inventory.statustab.confirmTitle")}
-        isOpen={isModalOpen}
-        onClose={handleModalToggle}
-        actions={[
-          <Button
-            key="confirm"
-            variant="primary"
-            data-testid={`${id}-set-state-modal-confirm`}
-            onClick={handleConfirm}
-          >
-            {words("yes")}
-          </Button>,
-          <Button
-            key="cancel"
-            variant="link"
-            data-testid={`${id}-set-state-modal-cancel`}
-            onClick={handleModalToggle}
-          >
-            {words("no")}
-          </Button>,
-        ]}
       >
-        {confirmationText}
-      </Modal>
-    </React.Fragment>
-  );
-};
-
-interface ToastAlertMessageProps {
-  stateErrorMessage: string;
-  id: string;
-  setStateErrorMessage: React.Dispatch<React.SetStateAction<string>>;
-}
-
-const ToastAlertMessage: React.FC<ToastAlertMessageProps> = ({
-  stateErrorMessage,
-  id,
-  setStateErrorMessage,
-}) => {
-  return (
-    <AlertGroup isToast={true}>
-      <Alert
-        variant="danger"
-        title={stateErrorMessage}
-        data-testid={`${id}-error-message`}
-        actionClose={
-          <AlertActionCloseButton
-            data-testid={`${id}-close-error-message`}
-            onClose={() => setStateErrorMessage("")}
-          />
-        }
-      />
-    </AlertGroup>
+        <Text> {confirmationText}</Text>
+      </ConfirmationModal>
+    </>
   );
 };
